adding blockchain_list_proposals Ticket #209

+++ libraries/blockchain/include/bts/blockchain/chain_database.hpp
@@ -107,9 +107,9 @@ namespace bts { namespace blockchain {
          /** top delegates by current vote, projected to be active in the next round */
          vector<account_id_type>            next_round_active_delegates()const;
                                             
-         vector<account_id_type>            get_delegates_by_vote( uint32_t first=0, uint32_t count = -1 )const;
-         vector<account_record>             get_delegate_records_by_vote( uint32_t first=0, uint32_t count = -1)const;
-         vector<proposal_record>            get_proposals( uint32_t first=0, uint32_t count = -1)const;
+         vector<account_id_type>            get_delegates_by_vote( uint32_t first=0, uint32_t count = uint32_t(-1) )const;
+         vector<account_record>             get_delegate_records_by_vote( uint32_t first=0, uint32_t count = uint32_t(-1))const;
+         vector<proposal_record>            get_proposals( uint32_t first=0, uint32_t count = uint32_t(-1))const;
          vector<proposal_vote>              get_proposal_votes( proposal_id_type proposal_id ) const;
 
          void                               scan_assets( function<void( const asset_record& )> callback );
